requirements-dev.txt
# Development-specific dependencies
pytest-xdist>=3.5
django-debug-toolbar>=4.0
ipython>=8.0
